requirements.txt
torch>=2.1.0
transformers>=4.45.0
datasets>=2.20.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
huggingface_hub>=0.24.0

openai>=1.40.0
python-dotenv>=1.0.0

scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.13.0
